Update distribution functions to prevent overflow on 32-bit systems

+++ Sources/BusinessMath/Statistics/Probability Distribution/erfInverse.swift
@@ -29,10 +29,10 @@ import Numerics
 /// Use this function when you need to perform a quantile function or percent-point function operation on your dataset.
 public func erfInv<T: Real>(y: T) -> T {
     let center = T(7) / T(10)
-    let a:[T] = [ T(0886226899) / T(1000000000), T(-1645349621) / T(1000000000),  T(0914624893) / T(1000000000), T(-0140543331) / T(1000000000)]
-    let b:[T] = [T(-2118377725) / T(1000000000),  T(1442710462) / T(1000000000), T(-0329097515) / T(1000000000),  T(0012229801) / T(1000000000)]
-    let c:[T] = [T(-1970840454) / T(1000000000), T(-1624906493 ) / T(1000000000),  T(3429567803) / T(1000000000),  T(1641345311) / T(1000000000)]
-    let d:[T] = [ T(3543889200) / T(1000000000),  T(1637067800) / T(1000000000)]
+	let a:[T] = [ T(Int(0886226899 as Double)) / T(Int(1000000000.0)), T(Int(-1645349621 as Double)) / T(Int(1000000000.0)),  T(Int(0914624893 as Double)) / T(Int(1000000000.0)), T(Int(-0140543331 as Double)) / T(Int(1000000000.0))]
+	let b:[T] = [T(Int(-2118377725 as Double)) / T(Int(1000000000.0)),  T(Int(1442710462 as Double)) / T(Int(1000000000.0)), T(Int(-0329097515 as Double)) / T(Int(1000000000.0)),  T(Int(0012229801 as Double)) / T(Int(1000000000.0))]
+	let c:[T] = [T(Int(-1970840454 as Double)) / T(Int(1000000000.0)), T(Int(-1624906493 as Double)) / T(Int(1000000000.0)),  T(Int(3429567803 as Double)) / T(Int(1000000000.0)),  T(Int(1641345311 as Double)) / T(Int(1000000000.0))]
+	let d:[T] = [ T(Int(3543889200 as Double)) / T(Int(1000000000.0)),  T(Int(1637067800 as Double)) / T(Int(1000000000.0))]
     if abs(y) <= center {
         let z = T.pow(y,2)
         let num = (((a[3]*z + a[2])*z + a[1])*z) + a[0]
